+++ ark/fuma/content/docs/definitions.mdx
@@ -8,7 +8,7 @@ import { SyntaxTab, SyntaxTabs } from "../../components/SyntaxTabs.tsx"
 
 ### Kitchen Sink
 
-```ts twoslash
+```ts
 export const currentTsSyntax = type({
 	keyword: "null",
 	stringLiteral: "'TS'",
@@ -37,7 +37,7 @@ export const currentTsSyntax = type({
 
 ## Constraints
 
-```ts twoslash
+```ts
 // runtime-specific syntax and builtin keywords with great error messages
 
 export const validationSyntax = type({
