+++ ark/fuma/content/docs/generics/index.mdx
@@ -8,7 +8,7 @@ sidebar:
 
 Record is now available as a builtin keyword.
 
-```ts twoslash
+```ts
 import { type } from "arktype"
 
 const stringRecord = type("Record<string, string>")
@@ -23,7 +23,7 @@ In addition to `Record`, the following generics from TS are now available in Ark
 
 These can be instantiated in one of three ways:
 
-```ts twoslash
+```ts
 import { type, type Type } from "arktype"
 
 const createBox = <T extends string>(of: type.Any<T>) =>
@@ -45,7 +45,7 @@ console.log(boxType({ box: "foo" }))
 
 #### Definition
 
-```ts twoslash
+```ts
 import { type } from "arktype"
 
 const boxOf = type("<t>", { box: "t" })
@@ -58,7 +58,7 @@ const schrodingersBox = boxOf({ cat: { isAlive: "boolean" } })
 
 All syntax in parameters definitions and all references to generic args are fully-type safe and autocompleted like any builtin keyword. Constraints can be used just like TS to limit what can be passed to a generic and allow that arg to be used with operators like `>`.
 
-```ts twoslash
+```ts
 import { type } from "arktype"
 
 const nonEmpty = type("<arr extends unknown[]>", "arr > 0")
@@ -70,7 +70,7 @@ const nonEmptyNumberArray = nonEmpty("number[]")
 
 There is a special syntax for specifying generics in a scope:
 
-```ts twoslash
+```ts
 import { scope } from "arktype"
 
 const types = scope({
@@ -85,7 +85,7 @@ const out = types.bitBox({ box: 0 })
 
 #### Invocation
 
-```ts twoslash
+```ts
 import { type } from "arktype"
 
 const one = type("Extract<0 | 1, 1>")
@@ -93,7 +93,7 @@ const one = type("Extract<0 | 1, 1>")
 
 ##### Chained
 
-```ts twoslash
+```ts
 import { type } from "arktype"
 
 const user = type({
@@ -108,7 +108,7 @@ const basicUser = user.pick("name", "age")
 
 #### Invoked
 
-```ts twoslash
+```ts
 import { type } from "arktype"
 
 const unfalse = type.keywords.Exclude("boolean", "false")
@@ -118,7 +118,7 @@ const unfalse = type.keywords.Exclude("boolean", "false")
 
 Our new generics have been built using a new method for integrating arbitrary external types as native ArkType generics! This opens up tons of possibilities for external integrations that would otherwise not be possible. As a preview, here's what the implementation of `Partial` looks like internally:
 
-```ts twoslash
+```ts
 import { generic, Hkt } from "arktype"
 
 const Partial = generic(["T", "object"])(
