+++ ark/fuma/content/docs/integrations/index.mdx
@@ -10,7 +10,7 @@ title: Integrations
 
 ArkType can easily be used with tRPC via the `assert` prop:
 
-```ts twoslash
+```ts
 // @noErrors
 t.procedure.input(
 	type({
