+++ ark/fuma/content/docs/intro/meta.json
@@ -1,0 +1,4 @@
+{
+	"title": "Intro",
+	"pages": ["setup", "morphs-and-more", "your-first-type", "adding-constraints"]
+}
