+++ ark/fuma/content/docs/primitives/meta.json
@@ -1,4 +1,12 @@
 {
 	"title": "Primitives",
-	"defaultOpen": true
+	"pages": [
+		"string",
+		"number",
+		"[bigint](/docs/primitives#bigint)",
+		"[symbol](/docs/primitives#symbol)",
+		"[boolean](/docs/primitives#boolean)",
+		"[null](/docs/primitives#null)",
+		"[undefined](/docs/primitives#undefined)"
+	]
 }
