+++ ark/fuma/content/docs/scopes/index.mdx
@@ -12,7 +12,7 @@ A scope is just like a scope in code- it's a resolution space where you can defi
 
 A lot of the time, if you don't need to create additional types in your scope, you can just `.export()` it right away.
 
-```ts twoslash
+```ts
 import { scope } from "arktype"
 
 // Scopes are collections of types that can reference each other.
